+++ benchmark/Speed.hs
@@ -1,26 +1,36 @@
+{-# LANGUAGE DerivingStrategies #-}
+{-# LANGUAGE ImportQualifiedPost #-}
+{-# LANGUAGE Trustworthy #-}
 {-# OPTIONS_GHC -Wall #-}
 
 module Main (main) where
 
-import Control.Monad.Bayes.Class
-import Control.Monad.Bayes.Inference.RMSMC
-import Control.Monad.Bayes.Inference.SMC
-import Control.Monad.Bayes.Population
-import Control.Monad.Bayes.Sampler
-import Control.Monad.Bayes.Traced
-import Control.Monad.Bayes.Weighted
+import Control.Monad.Bayes.Class (MonadInfer)
+import Control.Monad.Bayes.Inference.RMSMC (rmsmcLocal)
+import Control.Monad.Bayes.Inference.SMC (smcSystematic)
+import Control.Monad.Bayes.Population (runPopulation)
+import Control.Monad.Bayes.Sampler (SamplerIO, sampleIOwith)
+import Control.Monad.Bayes.Traced (mh)
+import Control.Monad.Bayes.Weighted (prior)
 import Criterion.Main
-import Criterion.Types
-import qualified HMM
-import qualified LDA
-import qualified LogReg
+  ( Benchmark,
+    Benchmarkable,
+    bench,
+    defaultConfig,
+    defaultMainWith,
+    nfIO,
+  )
+import Criterion.Types (Config (csvFile, rawDataFile))
+import HMM qualified
+import LDA qualified
+import LogReg qualified
 import System.Random.MWC (GenIO, createSystemRandom)
 
 -- | Environment to execute benchmarks in.
 newtype Env = Env {rng :: GenIO}
 
-data ProbProgSys = MonadInfer
-  deriving (Show)
+data ProbProgSys = MonadBayes
+  deriving stock (Show)
 
 data Model = LR [(Double, Bool)] | HMM [Double] | LDA [[String]]
 
@@ -59,7 +69,7 @@ prepareBenchmark e MonadInfer model alg =
   bench (show MonadInfer ++ sep ++ show model ++ sep ++ show alg) $
     prepareBenchmarkable (rng e) MonadInfer model alg
   where
-    sep = "_"
+    sep = "_" :: String
 
 -- | Checks if the requested benchmark is implemented.
 supported :: (ProbProgSys, Model, Alg) -> Bool
@@ -74,9 +84,9 @@ systems =
 lengthBenchmarks :: Env -> [(Double, Bool)] -> [Double] -> [[String]] -> [Benchmark]
 lengthBenchmarks e lrData hmmData ldaData = benchmarks
   where
-    lrLengths = 10 : map (* 100) [1 .. 10]
-    hmmLengths = 10 : map (* 100) [1 .. 10]
-    ldaLengths = 5 : map (* 50) [1 .. 10]
+    lrLengths = 10 : map (* 100) [1 :: Int .. 10]
+    hmmLengths = 10 : map (* 100) [1 :: Int .. 10]
+    ldaLengths = 5 : map (* 50) [1 :: Int .. 10]
     models =
       map (LR . (`take` lrData)) lrLengths
         ++ map (HMM . (`take` hmmData)) hmmLengths
@@ -98,9 +108,9 @@ lengthBenchmarks e lrData hmmData ldaData = benchmarks
 samplesBenchmarks :: Env -> [(Double, Bool)] -> [Double] -> [[String]] -> [Benchmark]
 samplesBenchmarks e lrData hmmData ldaData = benchmarks
   where
-    lrLengths = [50]
-    hmmLengths = [20]
-    ldaLengths = [10]
+    lrLengths = [50 :: Int]
+    hmmLengths = [20 :: Int]
+    ldaLengths = [10 :: Int]
     models =
       map (LR . (`take` lrData)) lrLengths
         ++ map (HMM . (`take` hmmData)) hmmLengths