+++ src/Control/Monad/Bayes/Inference/SMC2.hs
@@ -19,14 +19,17 @@ module Control.Monad.Bayes.Inference.SMC2
 where
 
 import Control.Monad.Bayes.Class
-    ( MonadCond(..), MonadInfer, MonadSample(random) )
-import Control.Monad.Bayes.Inference.RMSMC ( rmsmc )
-import Control.Monad.Bayes.Inference.SMC ( smcSystematicPush )
-import Control.Monad.Bayes.Population as Pop ( runPopulation, Population )
-import Control.Monad.Trans ( MonadTrans(..) )
-import Numeric.Log ( Log )
+  ( MonadCond (..),
+    MonadInfer,
+    MonadSample (random),
+  )
+import Control.Monad.Bayes.Inference.RMSMC (rmsmc)
+import Control.Monad.Bayes.Inference.SMC (smcSystematicPush)
+import Control.Monad.Bayes.Population as Pop (Population, runPopulation)
 import Control.Monad.Bayes.Sequential (Sequential)
 import Control.Monad.Bayes.Traced
+import Control.Monad.Trans (MonadTrans (..))
+import Numeric.Log (Log)
 
 -- | Helper monad transformer for preprocessing the model for 'smc2'.
 newtype SMC2 m a = SMC2 (Sequential (Traced (Population m)) a)