+++ src/Control/Monad/Bayes/Traced/Basic.hs
@@ -19,10 +19,18 @@ where
 
 import Control.Applicative (liftA2)
 import Control.Monad.Bayes.Class
-    ( MonadCond(..), MonadInfer, MonadSample(random) )
+  ( MonadCond (..),
+    MonadInfer,
+    MonadSample (random),
+  )
 import Control.Monad.Bayes.Free (FreeSampler)
 import Control.Monad.Bayes.Traced.Common
-    ( bind, mhTrans', scored, singleton, Trace(..) )
+  ( Trace (..),
+    bind,
+    mhTrans',
+    scored,
+    singleton,
+  )
 import Control.Monad.Bayes.Weighted (Weighted)
 import Data.Functor.Identity (Identity)
 import Data.List.NonEmpty as NE (NonEmpty ((:|)), toList)