+++ src/Control/Monad/Bayes/Traced/Static.hs
@@ -20,10 +20,14 @@ where
 
 import Control.Applicative (liftA2)
 import Control.Monad.Bayes.Class
-    ( MonadInfer, MonadCond(..), MonadSample(random) )
+  ( MonadCond (..),
+    MonadInfer,
+    MonadSample (random),
+  )
 import Control.Monad.Bayes.Free (FreeSampler)
 import Control.Monad.Bayes.Traced.Common
 import Control.Monad.Bayes.Weighted (Weighted, prior)
+
 import Control.Monad.Bayes.Weighted (Weighted)
 import Control.Monad.Trans (MonadTrans (..))
 import Data.List.NonEmpty as NE (NonEmpty ((:|)), toList)