+++ src/Control/Monad/Bayes/Weighted.hs
@@ -21,21 +21,21 @@ module Control.Monad.Bayes.Weighted
     prior,
     applyWeight,
     hoist,
-    toBinsWeighted
+    toBinsWeighted,
   )
 where
 
+import Control.Arrow (Arrow (first))
 import Control.Monad.Bayes.Class
   ( MonadCond (..),
     MonadInfer,
-    MonadSample (random),
+    MonadSample,
     factor
   )
 import Control.Monad.Trans (MonadIO, MonadTrans (..))
 import Control.Monad.Trans.State (StateT (..), mapStateT, modify)
-import Numeric.Log (Log (Exp))
+import Numeric.Log (Log)
 import Data.Fixed ( mod' )
-import Control.Arrow ( Arrow(first) )
 
 -- | Execute the program using the prior distribution, while accumulating likelihood.
 newtype Weighted m a = Weighted (StateT (Log Double) m a)
@@ -83,8 +83,5 @@ hoist t (Weighted m) = Weighted $ mapStateT t m
 
 toBinsWeighted :: Double -> [(Double, Log Double)] -> [(Double, Log Double)]
 toBinsWeighted binWidth = fmap (first (fst . toBin binWidth))
-  where toBin binSize n = let lb = n `mod'` binSize in (n-lb, n-lb + binSize) 
-
-
-
-
+  where
+    toBin binSize n = let lb = n `mod'` binSize in (n - lb, n - lb + binSize)