+++ test/Spec.hs
@@ -1,7 +1,11 @@
 
+{-# LANGUAGE LambdaCase #-}
+{-# LANGUAGE ImportQualifiedPost #-}
+{-# LANGUAGE Trustworthy #-}
+
 import Test.Hspec ( hspec, context, describe, it, shouldBe )
 import Test.Hspec.QuickCheck ( prop )
-import Test.QuickCheck ( Testable(property), (==>), ioProperty )
+import Test.QuickCheck ( (==>), ioProperty, property )
 import qualified TestEnumerator
 import qualified TestIntegrator
 import qualified TestInference
@@ -41,26 +45,24 @@ main = hspec do
         ls -> (TestIntegrator.volumeIsOne ls)
 
   describe "Integrator" $ do
-    it "gives correct answer for the sprinkler model" $
-      TestIntegrator.passed1 `shouldBe` True
-    it "computes expectation correctly" $
-      TestIntegrator.passed2 `shouldBe` True
-    it "gives correct answer for the sprinkler model" $
-      TestIntegrator.passed3 `shouldBe` True
-    it "computes expectation correctly" $
-      TestIntegrator.passed4 `shouldBe` True
-    it "gives correct answer for the sprinkler model" $
-      TestIntegrator.passed5 `shouldBe` True
-    it "computes expectation correctly" $
-      TestIntegrator.passed6 `shouldBe` True
-    it "gives correct answer for the sprinkler model" $
-      TestIntegrator.passed7 `shouldBe` True
-    it "computes expectation correctly" $
-      TestIntegrator.passed8 `shouldBe` True
-    it "gives correct answer for the sprinkler model" $
-      TestIntegrator.passed9 `shouldBe` True
-    it "gives correct answer for the sprinkler model" $
-      TestIntegrator.passed10 `shouldBe` True
+    it "" $ 
+      all (==True) [
+        TestIntegrator.passed1,
+        TestIntegrator.passed2,
+        TestIntegrator.passed3,
+        TestIntegrator.passed4,
+        TestIntegrator.passed5,
+        TestIntegrator.passed6,
+        TestIntegrator.passed7,
+        TestIntegrator.passed8,
+        TestIntegrator.passed9,
+        TestIntegrator.passed10,
+        TestIntegrator.passed11,
+        TestIntegrator.passed12,
+        TestIntegrator.passed13,
+        TestIntegrator.passed14
+        ] `shouldBe` True
+
   describe "Population" $ do
     context "controlling population" $ do
       it "preserves the population when not explicitly altered" $ do
@@ -119,6 +121,7 @@ main = hspec do
       prop "pipe model is equivalent to standard model" $
         \num -> property $ hmms $ take 5 num
     
+
   describe "SMC with stratified resampling" $
     prop "number of particles is equal to its second parameter" $
       \observations particles ->