+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 - No unreleased changes so far.
 
+# 0.1.1.1 (2021-06-08)
+
+Clean up of unused functions and broken code
+
+- remove unused functions in `Weighted` and `Population`
+- remove broken models in `models`
+- explicit imports
+- added some global language extensions
+
+
 # 0.1.1.0 (2020-04-08)
 
 - New exported function: `Control.Monad.Bayes.Class` now exports `discrete`.